fix(ios): CodeRabbit追加レビュー指摘対応
## Major修正
- Info.plist: NSHealthUpdateUsageDescriptionを削除（読み取り専用アプリのため不要）
- HealthKitRepository: 400行制限超過を解消
  - HealthKitRepository.swift (199行): 公開API + プロパティ + セーフラッパー
  - HealthKitRepository+Queries.swift (276行): プライベートクエリ実装
- LocationManager: String(localized:)での文字列補間を修正
  - %@プレースホルダーとString(format:)を使用

## Critical修正
- Date+Extensions.swift: 未使用ファイルを削除（14/15メソッドが未使用）

## Trivial修正
- LocalStorage: ISO8601DateFormatterをstatic letでキャッシュ化
